+++ Moveon/src/main/java/com/mphasis/moveon/serviceImpl/RouteServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import com.mphasis.moveon.daos.RouteDao;
 import com.mphasis.moveon.entities.Route;
 import com.mphasis.moveon.entities.Vehicle;
+import com.mphasis.moveon.exceptions.BusinessException;
 import com.mphasis.moveon.service.RouteService;
 @Service
 public class RouteServiceImpl implements RouteService {
@@ -29,8 +30,18 @@ public class RouteServiceImpl implements RouteService {
 
 	}
 
-	public void removeRoute(String route_Id) {
-		routeDao.deleteRoute(route_Id);
+	public Route removeRoute(String route_Id) throws BusinessException 
+	{
+		Route route = null;
+		if(route.getRoute_Id().equals(route_Id))
+		{
+			route = routeDao.deleteRoute(route_Id);
+		}
+		else
+		{
+			throw new BusinessException("Enter routeID is invalid" + route_Id);
+		}
+		return route;
 
 
 	}
