+++ Moveon/src/main/java/com/mphasis/moveon/serviceImpl/ScheduleServiceImpl.java
@@ -6,8 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.mphasis.moveon.daos.ScheduleDao;
+import com.mphasis.moveon.entities.Route;
 import com.mphasis.moveon.entities.Schedule;
 import com.mphasis.moveon.entities.Vehicle;
+import com.mphasis.moveon.exceptions.BusinessException;
 import com.mphasis.moveon.service.ScheduleService;
 
 @Service
@@ -26,9 +28,20 @@ public class ScheduleServiceImpl implements ScheduleService
 		scheduleDao.insertSchedule(schedule);
 	}
 
-	public void removeSchedule(String schedule_Id) 
+	public Schedule removeSchedule(String schedule_Id) throws BusinessException 
 	{
-		scheduleDao.deleteSchedule(schedule_Id);
+
+		Schedule schedule = null;
+		if(schedule.getSchedule_Id().equals(schedule_Id))
+		{
+			schedule = scheduleDao.deleteSchedule(schedule_Id);
+		}
+		else
+		{
+			throw new BusinessException("Enter scheduleID is invalid" + schedule_Id);
+		}
+		return schedule;
+
 	}
 
 	public List<Schedule> getAll()
